+++ R/omopcept-package.R
@@ -72,4 +72,5 @@ utils::globalVariables(c('concept_name', 'concept_id','concept_code',
                          'concept_id_1','concept_id_2',
                          'concept_name_1','concept_name_2',
                          'valid_start_date','valid_end_date','invalid_reason',
-                         'from', 'to', 'name', 'connections'))
+                         'from', 'to', 'name', 'connections',
+                         'concept_name_to_check','check'))