+++ NEWS.md
@@ -1,6 +1,7 @@
 
 # omopcept 0.0.5.6 2024-05-10 DEV version
 
+* `omop_cdm_combine()` refactored to make all IDs unique that it didn't before
 * `omop_relations()` gets `names2avoid` arg
 * `omop_graph()` gets args `nodealpha` `edgealpha` `edgewidth`